services/scheduler/pyproject.toml
[project]
name = "clawvis-scheduler"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.9",
  "apscheduler==3.10.4",
  "httpx==0.27.2",
  "pydantic==2.11.4",
  "pyyaml==6.0.2",
]

[dependency-groups]
dev = []
